fix migration data types

+++ database/migrations/2018_04_29_101754_create_user_product_table.php
@@ -14,10 +14,10 @@ class CreateUserProductTable extends Migration
     public function up()
     {
         Schema::create('user_product', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
         });
