+++ database/migrations/2018_04_29_121149_create_user_costumer_table.php
@@ -14,10 +14,10 @@ class CreateUserCostumerTable extends Migration
     public function up()
     {
         Schema::create('user_costumer', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('costumer_id')->unsigned();
+            $table->bigInteger('costumer_id')->unsigned();
             $table->foreign('costumer_id')->references('id')->on('costumers')->onDelete('cascade');
             $table->timestamps();
         });
