+++ database/migrations/2018_04_29_125309_create_user_supplier_table.php
@@ -14,10 +14,10 @@ class CreateUserSupplierTable extends Migration
     public function up()
     {
         Schema::create('user_supplier', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('supplier_id')->unsigned();
+            $table->bigInteger('supplier_id')->unsigned();
             $table->foreign('supplier_id')->references('id')->on('suppliers')->onDelete('cascade');
             $table->timestamps();
         });
